requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
bitarray==3.0.0
certifi==2024.8.30
charset-normalizer==3.4.0
cytoolz==1.0.0
Django==5.0
djangorestframework==3.14.0
eth-account==0.10.0
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-rlp==1.0.1
eth-typing==5.0.1
eth-utils==5.1.0
eth_abi==5.1.0
frozenlist==1.5.0
hexbytes==0.3.1
idna==3.10
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lru-dict==1.3.0
multidict==6.1.0
parsimonious==0.10.0
propcache==0.2.0
protobuf==5.28.3
pycryptodome==3.21.0
python-dotenv==1.0.0
pytz==2024.2
pyunormalize==16.0.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rlp==4.0.1
rpds-py==0.21.0
sqlparse==0.5.2
toolz==1.0.0
typing_extensions==4.12.2
urllib3==2.2.3
web3==6.11.1
websockets==14.1
yarl==1.18.0
Pillow==11.0.0
django-cors-headers==4.6.0